datadog: honor extracted sampling decisions (#30577)
honor extracted sampling decisions in the datadog tracer
Additional Description: See below.
Risk Level: low
Testing: See below.
Docs Changes: N/A
Release Notes (under bug fixes): honor extracted sampling decisions in the datadog tracer
Platform Specific Features: N/A

The Issue and Its Solution
Envoy users within Datadog noticed another bug caused by the migration of the Datadog tracer from OpenTracing to dd-trace-cpp that was released in Envoy 1.27.

Sometimes the Envoy spans are missing from the flame graphs of traces, and sometimes the entire subtree starting at the Envoy span is missing from the flame graphs of traces.

Datadog tracing in Envoy used to use the OpenTracing driver, but it has not since Envoy v1.27.

The OpenTracing driver sets the "sampling priority" to zero ("drop the trace") whenever bool Envoy::Tracing::Decision::traced is false. It does this regardless of whether a sampling decision was extracted from incoming request headers (e.g. x-b3-sampled or x-datadog-sampling-priority).

I followed suit in the non-OpenTracing Datadog tracer, but I missed something. The OpenTracing-based Datadog tracer, dd-opentracing-cpp, ignores sampling changes whenever a sampling decision has already been extracted from an incoming request.

So, in the OpenTracing-based code, when a sampling decision is extracted from an incoming request, here's what happened:

The trace's sampling decision is set to whatever was extracted.
If Envoy's sampling decision was "drop," then that decision is ignored even through the OpenTracing code tries to set the trace's decision accordingly.
The result is that extracted sampling decisions are always honored.

With the newer dd-trace-cpp based code, when a sampling decision is extracted from an incoming request, here's what happens:

The trace's sampling decision is set to whatever was extracted.
If Envoy's sampling decision was "drop," then the trace is dropped regardless of what was extracted.
So, if Envoy's configured sampling rate is less than 100%, users will see incomplete traces whenever Envoy is not the "root service." In some such traces, Envoy spans and possibly its descendants are missing.

This is a bug.

This revision restores the logic from the OpenTracing-based tracer: An Envoy decision of "drop" forces a dropped trace only if we haven't extracted a decision from an incoming request.

Testing
I added a unit test that checks a table of possiblities.

I also built Envoy and configured it to reverse proxy httpbin.org/headers with 1% sampling. Then I hit it with curl using a variety of propagation headers, such as x-datadog-sampling-priority and traceparent, and verified that the resulting sampling decision was consistent with what I sent.

When I don't send sampling information in the request headers, the resulting sampling decisions are consistent with Envoy's configured 1% rate.

Bug Fix
Please consider this PR as a bug fix so that it might be backported onto the v1.27 and v1.28 release branches.

Signed-off-by: David Goffredo <david.goffredo@datadoghq.com>