access log formatter: use new formatter context rather than multiple parameters (3/3) (#30758)
access log formatter: use new formatter context rather than multiple parameters (3/3)
Additional Description:

Continuous work of #30757. Waiting #30757 to be landed first.

The PR update the log Instance interface to be a specialization of InstanceBase with context type HttpFormatterContext.
And by this way, the log() method of Filter is updated to take a HttpFormatterContext reference as parameter.

This PR also do similar update to the AccessLogInstanceFactory and make it a specialization of AccessLogInstanceFactoryBase.

By this way, log instances of different modules will share the same code base.

Note, to ensure the backwards compatibility, the extension category of log instance should be kept as envoy.access_loggers.

This is the last PR of this big refactoring. After landed this PR, we will finally completely migrate our formatter/logger to use the context based interface.

This will bring two obvious benefits:

it's easier to make the common part of these code be a template. All formatter-based loggers could be shared by different modules like dubbo proxy, generic proxy or even redis/mysql if necessary (some minor update is still necessary, but it's simple and minor).
it's easier to extend the logger/formatter to add new feature like request trailer support. We needn't to change almost every log()/format() call to add a feature that only used in part of positions anymore. (think of the access log type 😥 ).
Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

Signed-off-by: wbpcode <wbphub@live.com>