dfp: Insert correct preresolved hostname key in DNS cache (#30784)
Previously, preresolved hostnames didn't include the port in the `host`
that is used as a key in the DNS cache. This behavior differed from the
regular resolution path which always normalized the host (via
DnsHostInfo::normalizeHostForDfp). There was also a bug in the
DnsCacheImplTest.PreresolveSuccess test, where the preresolve hostnames
config included the port in the socket address, which is not how it
would normally be specified and masked this issue.

This PR fixes the problem by ensuring the preresolved hostnames have the
host normalized (via DnsHostInfo::normalizeHostForDfp) before inserting
into the DNS cache, just like the host is normalized before retrieving
from the DNS cache. The test issue is also fixed so we can now verify
that the correct cache key is being used for insertion, retrieval, and
removal, and test coverage is added to ensure we can fetch the cache
entry correctly whether the host includes the port or not.

Signed-off-by: Ali Beyad <abeyad@google.com>