filter_manager_test: explicitly set the runtime flag no_downgrade_to_canonical_name to true in tests (#30855)
With #30340, the runtime flag no_downgrade_to_canonical_name is set to true in default. However, some downstream code have to maintain the legacy behavior for a while and have to set the default value to FALSE:

envoy/source/common/runtime/runtime_features.cc
Line 65 in 3c3372c

 RUNTIME_GUARD(envoy_reloadable_features_no_downgrade_to_canonical_name); 

If the default is false, some tests in filter_manager_test.cc can not pass since they are relying on the flag to be set to true. The solution is to explicitly set it into true in those tests to make them not depending on the default setting.

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>