Handle empty responses correctly in grpc_http1_reverse_bridge.
Previously, when an http 1 server returned a header-only response, the
grpc_http1_reverse_bridge bridge would fail to prepend a frame header,
because end_stream was ignored in encodeHeader, and encodeData is not
called for header-only responses.

That meant grpc clients were missing the expected frame header,
resulting in the request failing when it should succeed.

This commit ensures that empty responses have the grpc frame header
added.

As part of this change the FilterCallState::LastDataFrame bit is set on
the filter_call_state_ when calling encodeHeaders with end_stream=true,
because this is the first filter to need to call addDecodedTrailers
inside encodeHeaders.

This commit fixes the following issues:

https://github.com/envoyproxy/envoy/issues/29989

Signed-off-by: Henry Qin <hqin@squareup.com>