logging: Add has() methods for to HttpFormatterContext (#31022)
logging: Add has() methods for to HttpFormatterContext

HttpFormatterContext was recently added to store request headers, response headers and trailers. It provides simple accessors for those which return the static empty maps if not present. However, there are some envoy users which would like to test if the headers are present or not.

Signed-off-by: Ryan Hamilton <rch@google.com>