http3: Enable quic_reloadable_flag_quic_limit_sending_max_streams2 (#31132)
http3: Enable quic_reloadable_flag_quic_limit_sending_max_streams2

A recent QUICHE merge brough in this flag, but it caused DownstreamProtocols/ProtocolIntegrationTest.RetryStreamingCancelDueToBufferOverflow/IPv4_Http3Downstream_Http2UpstreamHttpParserNghttp2NoDeferredProcessingLegacy to crash, so the flag was forced to false. This happens because this flag has the side effect of delaying delivery of the STOP_SENDING frame to the test client, and so when the test completes, the decoder goes out of scope and is destroyed. Then when the code is destroyed, it tries to access the decoder and crashes.

This PR fixes the test by explicitly cleaning up before the test exits.

Risk Level: Low
Testing: Existing tests
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Ryan Hamilton <rch@google.com>