Use QuicheMemSlice constructor with releasor.
The existing model of forwarding to the quiche::MemSliceImpl
constructor is an abstraction violation and makes it hard to swap out
the underlying MemSlice platform implementation. As such, we've
recently added a new constructor to QuicheMemSlice (and the Impl API)
that allows for an arbitrary custom releasor.

This commit is a little wonky in that it only really uses a lambda
releasor for its capture list to manage object lifetime, and so the
lambda body is intentionally empty.

Since we're storing a unique_ptr in the capture list (allowable since
C++17), we can't store this in a std::function which requires
copyability. As such, we use absl::AnyInvocable.

Signed-off-by: Steve Wang <wangsteve@google.com>