mobile: changing the lifetime of k-v store (#31203)
previously the singleton nature of the PlatformInterfaceImpl meant a first envoy engine would setAPI to set the location of the platform bridge k-v store, we'd retreiveApi locate it and use it. A subsequent engine would setApi to a new location (for YT hopefully the same k-v store) but we'd never retreiveApi and use it. This meant for e2e tests where the value of the store changes across tests we never use the new location.

This fixes a failure mode in the H3 test where a second test pointed at the store for the first test.

Given that YT should have one k-v store location across multiple engines, this fix should ideally be a no-op. We could also address by resetting the singleton between tests.

Risk Level: medium
Testing: yes
Docs Changes: no
Release Notes: no

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>