Correct naming of  com_github_lyft_protoc_gen_star go_repository (#31169)
Similar to bufbuild/protoc-gen-validate#900
correct naming should include the major version.

https://go.dev/ref/mod#module-path

If the module is released at major version 2 or higher, the module path must end with a major version suffix like /v2. This may or may not be part of the subdirectory name. For example, the module with path golang.org/x/repo/sub/v2 could be in the /sub or /sub/v2 subdirectory of the repository golang.org/x/repo.

So go_repository stanza also should include the major version, to distinguish between the different major versions of the module, that can coexist.

Signed-off-by: Alex Bozhenko <abozhenko@pinterest.com>