protos: for Go, allow *opt-out* of PGV
PGV is extremely expensive in terms of binary size. This directly feeds
into memory usage, storage and transmission costs, etc. For our binary,
disable PGV (which we don't use) trims our binaries by a whopping 15MB!

This PR has no negative impact on any users unless they explicitly opt
out of PGV (with `-tags=disable_pgv`). All existing users are
un-impacted. This change only impacts Go via the go-control-plane repo.

This change is do as a post-processing step rather than in PGV upstream
as PGV has moved to v1.0 and totally rewritten the repo, so the old PGV
is frozen AFAIK. Since moving to v1.0 is a major effort (possibly not
planned), the simple approach here was taken.

This mirrors https://github.com/envoyproxy/envoy/pull/31172, where
similar proto additions are currently on hold due to the binary size
increase. There we are also planning to take a similar opt-in to the
binary size increase. In this case, we use an opt-out for backwards
compatibility, as PGV has been around longer.

Signed-off-by: John Howard <howardjohn@google.com>