mobile/ci: Update the mobile perf size constraints
A lot of great work has been done in the last few months to reduce the
Envoy Mobile binary size.  To prevent slippage, this commit readjusts
the maximum allowed test binary size in the Mobile/Perf CI job to
5.95MB, since the current size is about 5.907MB.

We also reduce the maximum percentage increase allowed to 1.5% (from
2%), to allow for catching binary size increases from PRs more
proactively.

Signed-off-by: Ali Beyad <abeyad@google.com>