api: Add trusted CIDRs to XFF config
X-Forwarded-For parsing can either trust a number of hops, as Envoy does
today, or it can trust by IP address.

This enables configuration to add a set of CIDRs to the XFF config, and
a recurse option.

When recurse is not enabled, HCM should trust a remote IP that is
covered by one of the configured CIDRs and use the last IP in XFF as the
remote IP.

When recurse is enabled, if the remote IP is covered by one of the
CIDRs, HCM should also walk backwards through XFF to find the first IP
not covered by the trusted CIDR list.

Signed-off-by: James O'Gorman <james@netinertia.co.uk>