http: Utility functions for parsing XFF by CIDR
remoteAddressIsTrustedProxy checks the remote IP against a list of
CIDRs. If it is covered by any CIDR the remote is considered a trusted
proxy.

getLastNonTrustedAddressFromXFF walks backwards through XFF to find the
first IP that is not covered by a list of trusted CIDRs. This IP is
considered the "true" remote IP.

Signed-off-by: James O'Gorman <james@netinertia.co.uk>