xff: Allow parsing XFF using trusted CIDRs
If a list of trusted CIDRs is set, rather than checking for a set number
of trusted hops, instead check the remote IP against the trusted CIDR
list.

If recurse is set, XFF is also evaluated to find the last IP in the list
that is not covered by the trusted CIDRs.

Signed-off-by: James O'Gorman <james@netinertia.co.uk>