Opentelemetry stream fix (#31844)
The OpenTelemetry exporter was previously using client-side streaming for exporting trace data. This appears to work fine with the reference OpenTelemetry collector, but some collectors (DataPrepper for one example) instead seem to expect unary RPC calls for trace exports (in accordance with the OpenTelemetry specification). It seems that Data Prepper appears to rely specifically on the end-stream flag arriving in order for it to properly process the trace exports.

Risk Level: medium
Testing: no new tests

Fixes #31497 

Signed-off-by: Ashish Banerjee <ashish.banerjee@solo.io>