http: scale adaptive concurrency gradient stat (#31734)
Gauge metrics are integers. Thus, when gradient is reported (which has values between 0.5 and 2), we end up only with 0, 1, or 2, which is of little-to-no value. This PR scales it by a 1000, which will give us values between 500 and 2000.

Risk Level: Minimal, unless users have setup systems that read this stat value and act upon it. This is doubtful though, since this metric has been practically useless.
Testing: n/a
Docs Changes: Clarified the stat values
Release Notes: Adaptive Concurrency Gradient stat was scaled by 1000 to improve its granularity
Platform Specific Features: N/A
Fixes #31695

Signed-off-by: Abdulrahman Alhamali <abdulrahman.alhamali@shopify.com>
Signed-off-by: AbdulRahmanAlHamali <a.alhamali93@gmail.com>