mobile: Fix a flake in the ClientIntegrationTest (#32177)
The ResetWithBidiTrafficExplicitData was crashing sporadically with
memory errors when run for HTTP3. Running with msan showed that the
crashes were caused in headers_callback.setReady() being called after
the test finished and member variables of the test were already
destroyed.

I noticed that in the test, we didn't call headers_callback.waitReady()
when the test was running for HTTP3.  It wasn't clear to me why, but
that seemed like the cause.  I removed the if-guard so we always call
headers_callback.waitReady(), and the test seems to pass now (over 500
runs).

Fixes #32024

Signed-off-by: Ali Beyad <abeyad@google.com>