network: Allow interfaceName() to gracefully handle getifaddrs failures
In the mobile world, we are finding getifaddrs returning ERRNO 19 on
occassion (ENODEV), causing a release assert to trigger and crash the
program. Given that interfaceName() is designed to return an optional
anyway, release assert is not the right thing to do here and instead, we
log a warning and return nullopt.

interfaceName() is mainly used for logging purposes, so this change
seems low risk.

Signed-off-by: Ali Beyad <abeyad@google.com>