Expose connection's streaminfo to the HTTP Lua filter
This allows Lua to access the connection StreamInfo, and thus the
dynamic metadata. This makes it possible for HTTP requests to act on
the Cloud provider specific PROXY PROTOCOL TLV extensions.

The bulk of the lines of this are moving the StreamInfoWrapper from
the Lua HTTP directory into the common Lua directory and fixing up the
references.

The rest of the lines are the, unfortunately, necessary de-consting of
Envoy::Network::Connection so that a mutable StreamInfo can provide a
mutable dynamic metadata, and the cascade of that up and around the
call stacks. This is not ideal, but seems better than using a
const_cast or forking StreamInfo in order to make the dynamic metadata
itself const.

Signed-off-by: Ryan Anderson <ryan@michonline.com>