util: add casting from OptRef<T> to OptRef<U> if U* = t_ptr is valid (e.g. adding 'const' and up-casting). (#32256)
Analagous to std::shared_ptr, you should be able to assign-construct OptRef<const T> from an OptRef<T> (adding const). This is possible to do now by spitting the declaration and assign, but a cast operator was needed to do it in one line. This PR simply adds that and a test. Without the new cast operator the test won't compile.

This cast also covers an assignment to the base type from a derived type.

Additional Description:
Risk Level: low
Testing: //test/...
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Joshua Marantz <jmarantz@google.com>