hcm: Support returning whether to append XFF
Allow original IP detection extensions to signal back to HCM whether it
should append to XFF. This currently only happens when
use_remote_address is true and extensions can't do it.

How the OriginalIPDetectionResult can signal if the extension should do
the same.

This allows extensions to work identically to when use_remote_address is
set. XFF now gains the ability to set this.

Signed-off-by: James O'Gorman <james@netinertia.co.uk>