mobile: Add the `no-remote-exec` tag to the Swift stats tests
For reasons we haven't figured out yet, RBE does not work well with the
Swift tests and causes frequent timeouts on CI.

See https://github.com/envoyproxy/envoy/issues/32551 for details.

We tagged the Swift tests with `no-remote-exec` but forgot to do so for
the swift/stats tests, so we're doing that here. This test timed out
recently on CI because we didn't have the `no-remote-exec` tag.

Signed-off-by: Ali Beyad <abeyad@google.com>