http3: fixing an upstream threading issue and bumping http3 upstream code back to alpha (#32640)
Envoy as a community is not in the habit of moving "GA" code to alpha. It was believed at the time that HTTP/3 upstream code had been deployed successfully in production but based on the severity of the bug fixed here it appears that was not the case. Changing security status to reflect the current perceived code stability.

Risk Level: low
Testing: e2e test
Docs Changes: inline
Release Notes: inline

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>