config: preserve presence of connect_config field
When converting a Route config proto into a RouteEntry, if there is an
UpgradeConfig with an upgrade_type of CONNECT or CONNECT-UDP, the
RouteEntry member variable connect_config_ is populated.

When this happens, the presence information associated with the
connect_config proto field is lost. That is, even if the connect_config
field was not present in the config proto, the connect_config_ member
variable will still be populated (with the default ConnectConfig value).

This is significant because, according to the documentation at
https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/http/upgrades#connect-support
the presence of the connect_config field should control whether CONNECT
requests are proxied to the upstream or terminated by Envoy.

To preserve this distinction, tweak the existing logic so that the
connect_config_ member variable is populated only when the
connect_config field is present.

Signed-off-by: Kenneth Jenkins <51246568+kenjenkins@users.noreply.github.com>