tests: fixed & expanded checking of ocsp response
In SslCertficateIntegrationTest.BothEcdsaAndRsaOnlyEcdsaOcspResponse,
the only check made on the OCSP response was on it's length not being
zero. However, in some error circumstances, the length given by
SSL_get0_ocsp_response() may be -1, which gives a false positive
result from the test. This commit expands the checking on the OCSP
response to check for the actual expected length and expected bytes.

Signed-off-by: Ted Poole <tpoole@redhat.com>