cluster: make dynamic updates to address-list provably thread-safe for LogicalHost, but not incur lock overhead for HostImpl. (#32919)
Commit Message: Resolves a race condition due to changing address-sets for LogicalHost during operation. The previous Host/HostDescription was hard to understand and prove correct from a thread safety perspective. In particular, it  returned a reference to a mutable structured via an API, and that is hard to protect from races. This PR resolves this by:

1. Returning a `shared_ptr<const vector<address>>` rather than a ref to a mutable vector from `addressList()`
2. Uses lock annotations to ensure member variables are safe during multi-threaded operations for LogicalHost
3. Uses `const` annotations to ensure member varaibles are  immutable and thus also threadsafe for HostImpl.
4. Adjusts class hierarchy to allow for maximum sharing of implementation between HostImpl and LogicalHost, but each one has its own member variables where necessary for the different thread safety models.

Additional Description:
WIthout the mutex, the new integraiton test fails under tsan 13/100 times. With the mutex it passes all the time. Note also that the mutex was there before this PR, but its use was not rigorously enforced by class structure or annotations. The fix was to change the API to return the address-list by shared_ptr rather than by reference.

Risk Level: medium -- this change to a central data-plane structure requires scrutiny. However the changes are mostly mechanical in order to enable the compiler to ensure thread safety for us.
Testing: //test/...
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a
Fixes: #32850

Signed-off-by: Joshua Marantz <jmarantz@google.com>