Expand worker id tag matcher to include thread_local_cluster_manager (#32744)
It was noticed in our local configuration (where we had a historical worker_id tag matcher) that our tag matcher conflicts with the default one and causes worker_id metrics not to be registered. However, there was one metric with a worker_id tag still being registered. My assumption is this was an oversight envoy-side, and that this one should be also picked up, which would make removing our custom tag matcher leave all the stats working correctly, rather than switching from thread_local_cluster_manager being the only one working, to thread_local_cluster_manager being the missing one.

Risk Level: Small, change is only to sunk metrics.
Testing: Same as it is for the rest of the metrics in this regex.
Docs Changes: n/a
Release Notes: Described the change.
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>