mobile: Adds support for setting network thread priority in the C++ API
A new method, `setNetworkThreadPriority`, is added to the EngineBuilder,
which enables the user to set the thread priority of the Envoy Mobile
engine thread.

The thread priority is set using the setpriority() system call, which
takes in "nice" values in the range of -20 to 19, where -20 is the
highest priority and 19 is the lowest priority.

Signed-off-by: Ali Beyad <abeyad@google.com>