Add another #ifdef guard for QUIC.
In previous commits, I added a few other #ifdef guards to remove
getCryptoConfig() if we're not building with QUIC, but it seems that
there was another test depending on this method existing.

I grepped through the codebase for other instances of
getCryptoConfig() and this seems like the last instance that was
unguarded and might be built with QUIC disabled (the other cases
involved QUIC-specific code).

Signed-off-by: Steve Wang <wangsteve@google.com>