getaddrinfo: Treat EAI_NONAME and EAI_NODATA as non-failures
When the DNS resolution status is Failure, the DNS cache treats it as a
retryable error, using a retry policy to enable a timer for retries.
With EAI_NONAME and EAI_NODATA, we do not expect these to be transient
failures, so the getaddrinfo now treats it as a succesful query with
empty results, so we don't trigger the failure refresh/retry timer.

NOTE: The c-ares resolver also treats NODATA and NONAME results as
successful queries.

Signed-off-by: Ali Beyad <abeyad@google.com>