mobile: Fix the test_binary_size program to use the InternalEngine (#34010)
The previous test_binary_size code was:
reinterpret_cast<Envoy::InternalEngine*>(0)->run(nullptr, nullptr);

This code has a couple problems: (1) we don't actually create an InternalEngine; we just cast an arbitrary
integer value (in this case, 0) to a InternalEngine pointer and (2) the run() function takes const references,
not pointers. The way this code was structured, it's believed that the binary size expectations were incorrect,
because the compiler could strip/elide code that this particular program setup would render unused code.

After changing the program structure, we do indeed see a size difference. Instead of 4.5MB, as is the current
expectation, we are seeing the program size actually as 5.41MB.

Signed-off-by: Ali Beyad <abeyad@google.com>