mobile: Refactor the stream callback implementation (#34014)
This PR refactors the stream callback implementation by removing the envoy_http_callbacks C wrapper and Envoy::Platform::StreamCallbacks. The new implementation uses the built-in C++ Envoy Core data types as much as possible instead of going through the wrapper types. By removing the C wrapper types, we get the benefit of not having to deal with manual memory management. Prior to this PR, the Envoy::Platform::StreamCallbacks was a shared_ptr. By changing the StreamPrototype::start to take a move-only EnvoyStreamCallbacks, we no longer need to use a shared_ptr, which IMO is a better design.

The Android and iOS implementations still use some C wrapper types because to getting rid of the C wrapper types completely requires a lot of changes on both Android and iOS implementation, which is best done in a follow up PR.

Risk Level: medium
Testing: unit & integration tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: mobile

Signed-off-by: Fredy Wijaya <fredyw@google.com>