mobile: Add respones codes to the error details
The HTTP response code is not reported in the Cronvoy exceptions, and
the error_code, which maps from some small subset of response codes to
envoy_error_code_t, is not used in the onError functions. As a first
step, we want to append the HTTP response code and the Envoy error code
to the error details.

A subsequent PR will add more support for mapping Http::Code to
envoy_error_code_t (or getting rid of envoy_error_code_t all together).

Signed-off-by: Ali Beyad <abeyad@google.com>