Synchronise expiry of id token cookie with JWT exp claim
The expiry of the id token cookie should match the exp claim in the id
token JWT, not the expiry of the access token. Some identity providers
(definitely Microsoft Entra ID, probably others) issue id tokens with
different expiry than the access tokens.

This change is largely based on the work done in #32278 which supports expiry
of the refresh token based on the exp claim of the refresh token if it
is a JWT.

Closes #32485.

Signed-off-by: Michael Sauter <mail@michaelsauter.net>